Fix delete word not deleting newlines; properly update max scroll

+++ gui/renderer/atlas.cc
@@ -31,7 +31,7 @@ Atlas::Atlas() {
         // TODO: Creating this` vector is quite slow, so disable during release.
         atlas_background = std::vector<uint8_t>(kAtlasSize * kAtlasSize * 4, 0);
         size_t pixels = kAtlasSize * kAtlasSize;
-        for (int i = 0; i < pixels; ++i) {
+        for (size_t i = 0; i < pixels; ++i) {
             size_t offset = i * 4;
             atlas_background[offset] = r;
             atlas_background[offset + 1] = g;
